Change backend after code-review
- no logic in the api, more functions in the modules
- new module streetview
- implement BaseModel
- implement tracing
