+++ guess-the-country/country-backend/country/explainer/explain.py
@@ -3,10 +3,32 @@ from skimage.segmentation import mark_boundaries
 import base64
 import cv2
 import numpy as np
-from ..model.predict import model
+from ..model.predict import model, load_image, preprocess
+from pydantic import BaseModel
+import uuid
+from xaidemo.tracing import traced
+
+class Explanation(BaseModel):
+    explanation_id: uuid.UUID
+    image: bytes
+
+@traced
+def explain(data):
+    encoded_data = str(data)
+    image = load_image(encoded_data)
+    pre_image = preprocess(image)
+    explanation = explain_cnn(pre_image, model)
+    explain_id = uuid.uuid4()
+    encoded_image_string = convert_explanation(explanation)
+    encoded_bytes = bytes("data:image/png;base64,",
+                          encoding="utf-8") + encoded_image_string
+    return Explanation(
+        explanation_id=explain_id,
+        image=encoded_bytes
+    )
 
 
-
+@traced
 def convert_explanation(explanation):
     image = (explanation*255).astype(np.uint8)
     image_rgb = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
@@ -16,8 +38,8 @@ def convert_explanation(explanation):
     encoded_image_string = base64.b64encode(buffer)
     return encoded_image_string
 
-
-def explain_cnn(img):
+@traced
+def explain_cnn(img, model):
     explainer = lime_image.LimeImageExplainer()
     #explanation = explainer.explain_instance(img[0].astype('double'), model.predict,  segmentation_fn =  SegmentationAlgorithm('slic', kernel_size=4, max_dist=100, ratio=0.4), top_labels=2, hide_color=False, num_samples=1000)
     explanation = explainer.explain_instance(img[0].astype(
