+++ guess-the-country/country-backend/requirements.txt
@@ -10,5 +10,6 @@ scikit-image
 numpy
 tensorflow-cpu==2.7.0
 python-multipart
+pydantic
 
 git+https://github.com/xai-demonstrator/xai-demonstrator.git@318ad17a9ebf2760f4f23b45cfdc29d18d393ec4#egg=xaidemo&subdirectory=common/backend-utils
